Recover 31 first-party functions and typed ABI boundaries
Recover save, script, battle, field, calendar, and drawing paths. Replace synthetic renderer parameters with genuine packed variadics, migrate task and halfword contracts, and register the required GP symbols.

Verify 6213 first-party MATCH / 647 ASM, 172 source objects, both retail SHA-1s, 529 passing tests, and clean source-honesty lint.

+++ src/Kosaka/k_command/k_command.c
@@ -748,7 +748,7 @@ u32 func_00178970()
 // FUN_0025D100
 u32 func_0025d100()
 {
-    func_0029cf50(func_00106600(func_0029cc00(0)));
+    func_0029cf50((u8)func_00106600((s16)func_0029cc00(0)));
 
     return true;
 }
